refactor(cf-chat): migrate to new multi-store API
