+++ examples/cf-chat/src/livestore/store.ts
@@ -1,0 +1,34 @@
+import { makePersistedAdapter } from '@livestore/adapter-web'
+import LiveStoreSharedWorker from '@livestore/adapter-web/shared-worker?sharedworker'
+import { useStore } from '@livestore/react'
+import { unstable_batchedUpdates as batchUpdates } from 'react-dom'
+import LiveStoreWorker from '../livestore.worker.ts?worker'
+import { schema } from './schema.ts'
+
+const adapter = makePersistedAdapter({
+  storage: { type: 'opfs' },
+  worker: LiveStoreWorker,
+  sharedWorker: LiveStoreSharedWorker,
+})
+
+const getStoreId = () => {
+  if (typeof window === 'undefined') return 'unused'
+
+  const searchParams = new URLSearchParams(window.location.search)
+  const storeId = searchParams.get('storeId')
+  if (storeId !== null) return storeId
+
+  const newAppId = crypto.randomUUID()
+  searchParams.set('storeId', newAppId)
+
+  window.location.search = searchParams.toString()
+  return newAppId
+}
+
+export const useAppStore = () =>
+  useStore({
+    storeId: getStoreId(),
+    schema,
+    adapter,
+    batchUpdates,
+  })
